Adding globalblock-exempt grant for OAuth
Accounts with the global IP block exemption flag operating via OAuth
consumers, such as pywikibots but not only those, are still being
affected by global IP blocks because their user rights ain't recognized.
This patch adds the globalblock-exempt right to the basic grants list,
so that if the user has it via any explicit or implicit user group,
it can make use of it.

Bug: T142306
Change-Id: I04b6dcf38daaac61ab373e9ceece4716d3f7a4b1