Use HTMLForm in SpecialGlobalBlockList
The error message is now shown below the form and is
also now shown as a single error message instead of as
a list (which didn't really make sense because there is
only one error message which can be shown on this page
currently). This also removes the 'globalblocking-unblock-errors'
message which was being wrongly used on this page.

Change-Id: I454e6b889d47a01ef8d7e961a9ba7925915cc955