Link to Special:GlobalBlock instead of Special:GlobalBlockList on block success
A link to Special:GlobalBlockList is already available on the subtitle
so this is kind of redundant. Additionally, sometimes users would need
to do another block after doing a block and pressing the browser "Back"
button isn't very reliable since this is a POST form. To ease that workflow
a little, add a handy link here.

Change-Id: Ib2b2ab23d3cc7f826bcf9d40762cd7fb09cbed80