Update for API error i18n
See Iae0e2ce3.

To allow for message reuse, this mostly copies the CIDR-handling logic
from ApiQueryBlocks into ApiQueryGlobalBlocks, which has the side effect
of making it actually work for IPv6. Too much refactoring of the rest of
the code would be needed to properly solve T46836 for me to do right
now, though.

Change-Id: I53987bf87c48f6c00deec17a8e957d24fcc3eaa6