Update for Block::setBlocker() requiring a valid user or interwiki username
In prepraration for the upcoming actor table, Block requires the blockers
to either be a local user or an interwiki username.

We can use CentralIdLookup to check whether the blocker on the foreign wiki
is the same as the local wiki, otherwise fallback to an interwiki username
that uses the wiki as a prefix.

Bug: T182344
Change-Id: Ia3c85fced974a4edabc172bd3ef6a7a255ffa462