Set block parameters on GlobalBlock object to fix error message
Block errors are displayed by the BlockErrorFormatter and API users
by building messages from the block parameters. However before this,
the GlobalBlock object does not set these parameters, instead
expecting a customized error message to be displayed in full, leading
to an incomplete message being displayed on mobile.

This sets the parameters, to allow an error message to be built from
a GlobalBlock via the API and the BlockErrorFormatter.

Bug: T233996
Change-Id: I9d2127532e486d293318f4e25e2c1ebc66a0c05b