Return early from GetUserBlock hook handler if no IP to check
If no IP address is passed to the handler, and the passed user name is
not an IP either, then there are no blocks for us to find, so return
early.

Without this early return, we may accidentally find a block against the
global session user, when we're not looking for one.

Bug: T326256
Change-Id: Ie8dad4be1b8e8f72c6ea9e2876e34ff90c633098