Fix test that only works in June or July
The test created a block with an expiry time of 1 month, and asserted
that the formatted row contains "july". There was no attempt to mock the
time. The comment which says that the "july" message is only used in the
expiry is incorrect -- it is also used in the timestamp. That's why the
test only started failing in August.

So, set the expiry to a specific date a long way in the future. Check
for that date.

Bug: T371577
Change-Id: I024c59574d1a1c08747d1bacffb80998d2b11563
(cherry picked from commit 32b5f7ea7c7cb03c4d00e7ef02ef7fe744dbbf05)