GlobalBlockLookup: Fix getGlobalBlockId() returning expired blocks
Right now, methods calling getGlobalBlockId(), including the
CentralAuthHooks code resposible for adding information about global
blocks to Special:CentralAuth, would receive a block ID if there was an
expired block still in the database.

Upon a cursory glance of its callers, they all use the method in a way
that expects active blocks only, so change the method itself to behave
in a more obvious way instead of touching the callers.

Change-Id: Id1bc9807c021859edeb9d7cf5ddc940fe31147ee