Don't run fieldExists() from LoadExtensionSchemaUpdates
LoadExtensionSchemaUpdates is meant to be purely declarative, you're not
meant to actually do things when the hook is called. That includes
getting connections. There's no point checking for field existence here
since we may not have created the tables yet when the hook is called.

The check for gb_by existence is trying to detect whether it is safe to
run the patch on SQLite, but this is not completely successful. Really
we want to make it conditional on non-existence of gb_target_central_id.
So implement a helper in core and use that instead.

Move the isNullable() check to the maintenance script. Don't use
GlobalBlockingServices since extension services have not been
registered when this hook is run from install.php. Apply the temporary
schema on all DB types so that isNullable() returns the correct value.

Use modifyFieldIfNullable, introduced for this purpose, to make
gb_autoblock_parent_id not nullable.

This fixes an exception from the proposed installPreConfigured.php with
the shared database creation feature. fieldExists() tries to connect to
the shared database before virtual domain database creations are done,
and so throws a connection error.

Depends-On: I61910dba894c164bebb880aa6af39cc534bd64c5
Change-Id: I50facb7823a971725e109da8a26d738facc55b57